+++ src/main/kotlin/com/bee/newsfeed_bee/domain/reply/controller/ReplyController.kt
@@ -1,10 +1,10 @@
-package com.bee.newsfeed_bee.reply.controller
+package com.bee.newsfeed_bee.domain.reply.controller
 
-import com.bee.newsfeed_bee.reply.dto.CreateReplyRequest
-import com.bee.newsfeed_bee.reply.dto.DeleteReplyRequest
-import com.bee.newsfeed_bee.reply.dto.ReplyResponse
-import com.bee.newsfeed_bee.reply.dto.UpdateReplyRequest
-import com.bee.newsfeed_bee.reply.service.ReplyService
+import com.bee.newsfeed_bee.domain.reply.dto.CreateReplyRequest
+import com.bee.newsfeed_bee.domain.reply.dto.DeleteReplyRequest
+import com.bee.newsfeed_bee.domain.reply.dto.ReplyResponse
+import com.bee.newsfeed_bee.domain.reply.dto.UpdateReplyRequest
+import com.bee.newsfeed_bee.domain.reply.service.ReplyService
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
@@ -34,19 +34,19 @@ class ReplyController(
     }
 
     @PostMapping()
-    fun createReply(@PathVariable feedId:Long, request:CreateReplyRequest ) : ResponseEntity<ReplyResponse>
+    fun createReply(@PathVariable feedId:Long, request: CreateReplyRequest) : ResponseEntity<ReplyResponse>
     {
         return ResponseEntity.status(HttpStatus.CREATED).body(replyService.createReply(feedId, request))
     }
 
     @PutMapping("/{replyId}")
-    fun updateReply(@PathVariable feedId: Long, @PathVariable replyId: Long, request:UpdateReplyRequest) : ResponseEntity<ReplyResponse>
+    fun updateReply(@PathVariable feedId: Long, @PathVariable replyId: Long, request: UpdateReplyRequest) : ResponseEntity<ReplyResponse>
     {
         return ResponseEntity.status(HttpStatus.OK).body(replyService.updateReply(replyId, request))
     }
 
     @DeleteMapping("/{replyId}")
-    fun deleteReply(@PathVariable feedId: Long, @PathVariable replyId: Long, request:DeleteReplyRequest) : ResponseEntity<Unit>
+    fun deleteReply(@PathVariable feedId: Long, @PathVariable replyId: Long, request: DeleteReplyRequest) : ResponseEntity<Unit>
     {
         replyService.deleteReply(feedId, replyId, request)
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build()
