+++ src/main/kotlin/com/bee/newsfeed_bee/domain/reply/dto/CreateReplyRequest.kt
@@ -1,7 +1,7 @@
-package com.bee.newsfeed_bee.reply.dto
+package com.bee.newsfeed_bee.domain.reply.dto
 
 data class CreateReplyRequest (
     val userName:String,
     val password:String,
-    val content:String
+    val contents:String
 )
