+++ src/main/kotlin/com/bee/newsfeed_bee/domain/reply/dto/DeleteReplyRequest.kt
@@ -1,4 +1,4 @@
-package com.bee.newsfeed_bee.reply.dto
+package com.bee.newsfeed_bee.domain.reply.dto
 
 data class DeleteReplyRequest (
     val password:String
