+++ src/main/kotlin/com/bee/newsfeed_bee/domain/reply/dto/UpdateReplyRequest.kt
@@ -1,0 +1,6 @@
+package com.bee.newsfeed_bee.domain.reply.dto
+
+data class UpdateReplyRequest (
+    val password:String,
+    val contents:String
+)
