--- src/main/kotlin/com/bee/newsfeed_bee/reply/dto/UpdateReplyRequest.kt
@@ -1,6 +1,0 @@
-package com.bee.newsfeed_bee.reply.dto
-
-data class UpdateReplyRequest (
-    val password:String,
-    val content:String
-)
